Вынос ConsolePrintWriter из неработающего интерфейса
